Refatorando o componente de cadastro de usuário para utilizar componentes de entrada separados para email e perfis.
