+++ front/src/components/cadastro_usuario/styles.module.css
@@ -3,70 +3,3 @@
   flex-direction: column;
   gap: 12px;
 }
-
-.CadastroUsuario label {
-  font-weight: bold;
-  margin-bottom: -8px;
-}
-
-.CadastroUsuario input,
-.CadastroUsuario select {
-  padding: 8px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-size: 14px;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.CadastroUsuario .perfilSelect {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    width: 100%;
-}
-
-.CadastroUsuario .perfilSelect select {
-    flex: 1 1 auto;
-    min-width: 0;
-}
-
-.CadastroUsuario .perfilSelect button {
-    flex: 0 0 120px;
-    padding: 8px 12px;
-    cursor: pointer;
-    background-color: silver;
-    border: none;
-    border-radius: 5px;
-    font-weight: normal;
-    color: #333;
-    font-size: 16px;
-    align-items: center;
-    margin-left: auto;
-}
-
-.CadastroUsuario .perfis {
-    display: flex;
-    gap: 8px;
-    flex-wrap: wrap;
-}
-
-.CadastroUsuario .perfis span {
-    vertical-align: center;
-    padding: 4px 8px;
-    background: #eee;
-    border-radius: 4px;
-}
-
-.CadastroUsuario .clearBtn {
-    border: none;
-    background: transparent;
-    color: #888888;
-    font-size: 16px;
-    margin-left: 14px;
-    cursor: pointer;
-}
-
-.CadastroUsuario .clearBtn:hover {
-    color: #555555;
-}
