+++ front/src/components/input_email/styles.module.css
@@ -1,0 +1,13 @@
+.label {
+    font-weight: bold;
+    margin-bottom: -8px;
+}
+
+.input {
+    padding: 8px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 14px;
+    width: 100%;
+    box-sizing: border-box;
+}
