+++ front/src/components/input_perfis/styles.module.css
@@ -1,0 +1,69 @@
+.label {
+    font-weight: bold;
+    margin-bottom: -8px;
+}
+
+.select {
+    padding: 8px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 14px;
+    width: 100%;
+    box-sizing: border-box;
+}
+
+.perfilSelect {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    width: 100%;
+}
+
+.perfilSelect select {
+    flex: 1 1 auto;
+    min-width: 0;
+}
+
+.perfilSelect button {
+    flex: 0 0 120px;
+    padding: 8px 12px;
+    cursor: pointer;
+    background-color: silver;
+    border: none;
+    border-radius: 5px;
+    font-weight: normal;
+    color: #333;
+    font-size: 16px;
+    align-items: center;
+    margin-left: auto;
+}
+
+.perfilSelect button:hover {
+    background-color: #e0e0e0;
+}
+
+.perfis {
+    display: flex;
+    gap: 8px;
+    flex-wrap: wrap;
+}
+
+.perfis span {
+    vertical-align: center;
+    padding: 4px 8px;
+    background: #eee;
+    border-radius: 4px;
+}
+
+.clearBtn {
+    border: none;
+    background: transparent;
+    color: #888888;
+    font-size: 16px;
+    margin-left: 14px;
+    cursor: pointer;
+}
+
+.clearBtn:hover {
+    color: #555555;
+}
